26-03-04

+++ charang/frontend/src/pages/Manager/AllUserPage.jsx
@@ -5,15 +5,17 @@ import { useNavigate } from "react-router-dom";
 import './AllUserPage.css';
 import axios from "axios";
 
-export default function AllUserPage(){
+export default function AllUserPage() {
 
-    const {pagesHandler, paging, pageNum, setPageNum, user,
-        userFind, userSearchType, setUserSearchType, setUserSearchWord, bookStatusFind, allBookStatus} = useContext(DataContext);
-        // --- 추가된 초기화 로직 ---
+    const { pagesHandler, paging, pageNum, setPageNum, user, searchResetHandler,
+        userFind, userSearchType, setUserSearchType, setUserSearchWord, bookStatusFind, allBookStatus } = useContext(DataContext);
+    // --- 추가된 초기화 로직 ---
     useEffect(() => {
         // 다른 페이지에서 진입 시 무조건 1페이지로 시작
         setPageNum(1);
         setUserSearchWord('');
+        searchResetHandler();
+        userCount();
     }, []); // 빈 배열([])을 넣어 마운트 시점에 딱 한 번만 실행되게 합니다.
 
     // 회원 검색 핸들러
@@ -23,10 +25,11 @@ export default function AllUserPage(){
     }
     // -----------------------
     // 전체 예약, 전체 회원 출력 함수 호출
-     useEffect(()=>{
+    useEffect(() => {
         userFind();
         bookStatusFind();
-    },[pageNum]);
+        userCount();
+    }, [pageNum]);
 
 
     // 회원 삭제
@@ -37,48 +40,48 @@ export default function AllUserPage(){
         let delUserCopy = [...delUser];
 
         // 체크되어있으면 delUser배열에 넣기(true)
-        if(e.target.checked){
+        if (e.target.checked) {
             delUserCopy.push(userId);
             setDelUser(delUserCopy);
         }
         // 체크를 했다가 취소할 경우(false)
-        else{
+        else {
             delUserCopy = delUser.filter(id => id !== userId);
             setDelUser(delUserCopy);
         }
     }
 
     const delHandler = () => {
-        if(delUser.length == 0){
-            alert("삭제할 회원을 선택해주세요.");
+        if (delUser.length == 0) {
+            alert("삭제할 예약을 선택해주세요.");
             return;
-        }else{
-            axios.delete("/api/delete",{data:delUser})
-            .then((res)=>{
-                console.log("삭제 결과: ", res.data);
-                if(res.data == 1){
-                    alert("삭제되었습니다");
-                    userFind();
-                }else{
-                    alert("다시 시도해주세요.");
-                }
-            })
-            .catch((error)=>{
-                console.log("받아온 삭제 결과 에러: ", error);
-            })
+        } else {
+            const confirmCancel = window.confirm(`${delUser.length}명의 회원데이터를 삭제하시겠습니까?`);
+            if (!confirmCancel) return;
+            axios.delete("/api/delete", { data: delUser })
+                .then((res) => {
+                    console.log("삭제 결과: ", res.data);
+                    if (res.data == 1) {
+                        alert(`${delUser.length}명의 회원데이터가 삭제되었습니다`);
+                        userFind();
+                        setDelUser([]);
+                        userCount();
+                    } else {
+                        alert("다시 시도해주세요.");
+                    }
+                })
+                .catch((error) => {
+                    console.log("받아온 삭제 결과 에러: ", error);
+                })
         }
     }
-
 
     // 예약이 앞으로 존재하는 회원인지 구분
     const noRes = (userId) => {
         // 앞으로 예약이 존재하는 회원
         const ingUser = allBookStatus
-        .filter(pastStatus => pastStatus.bookingStatus === "UPCOMING" || pastStatus.bookingStatus === "ONGOING")
-        .map(res=>res.userId);
-        // console.log("전체 예약 회원: ",allBookStatus);
-        // console.log("삭제 불가능 회원: ",ingUser);
-        // console.log(`현재 페이지 전체 예약수: ${allBookStatus.length}건 / 삭제불가 회원수: ${ingUser.length}명`);
+            .filter(pastStatus => pastStatus.bookingStatus === "UPCOMING" || pastStatus.bookingStatus === "ONGOING")
+            .map(res => res.userId);
         return ingUser.includes(userId);
     }
 
@@ -92,15 +95,27 @@ export default function AllUserPage(){
     // placeholder
     const placeholderWord = () => {
         console.log("검색", userSearchType);
-         if(userSearchType === 'userId'){
+        if (userSearchType === 'userId') {
             return "아이디를 검색하세요";
-        }else{
+        } else {
             return "이름을 검색하세요";
         }
     }
 
+    // 전체 회원 개수 불러오기
+    const [userCnt, setUserCnt] = useState(0);
+    const userCount = () => {
+        axios.get("/api/allUserCount")
+            .then((res) => {
+                console.log("받아온 전체 회원 개수: ", res.data);
+                setUserCnt(res.data);
+            })
+            .catch((error) => {
+                console.log("불러온 전체 회원 개수 에러: ", error);
+            })
+    }
 
-    return(
+   return(
         <div className="ManagerAllUser">
             <h1>전체 회원목록</h1>
 
@@ -110,7 +125,7 @@ export default function AllUserPage(){
                 <select name="userSearchType" className="mau_select"
                 onChange={(e)=> setUserSearchType(e.target.value)}>
                     <option value="userId">회원ID</option>
-                    <option value="userName">회원이름</option>
+                    <option value="model">회원이름</option>
                 </select>
                 {/* 검색 단어*/}
                 <input type="text" name="searchWord" className="mau_input" placeholder={placeholderWord()}
@@ -133,7 +148,7 @@ export default function AllUserPage(){
                         <th className="managerAllUser_userResiNum">주민등록번호</th>
                         <th className="managerAllUser_userPhone">휴대폰번호</th>
                         <th className="managerAllUser_userRegDate">가입일자</th>
-                        <th className="managerAllUser_userDel">회원삭제</th>
+                        <th className="managerAllUser_userDel">삭제<p>({delUser.length}/{userCnt})</p></th>
                     </tr>
                 </thead>
                 {user && user.length > 0 ? 
@@ -158,7 +173,7 @@ export default function AllUserPage(){
                                     {noRes(user.userId)?
                                         <p>불가</p>
                                         :
-                                        <input type="checkbox" name="delcheck" 
+                                        <input type="checkbox" name="delcheck" checked={delUser.includes(user.userId)}
                                         className="AllUser_del" onChange={(e)=>{checkHandler(e, user.userId)}}></input>
                                     }
